chore: remove chapi bridge
- chapi bridge is moved to edge adapters as wallet bridge
- Closes #651
