+++ cmd/wallet-server/startcmd/start.go
@@ -81,11 +81,6 @@ const (
 		" Alternatively, this can be set with the following environment variable: " + dependencyMaxRetriesFlagEnvKey
 	dependencyMaxRetriesDefault = uint64(120) // nolint:gomnd // false positive ("magic number")
 
-	walletAppURLFlagName  = "wallet-app-url"
-	walletAppURLFlagUsage = "Wallet App URL." +
-		" Alternatively, this can be set with the following environment variable: " + walletAppURLEnvKey
-	walletAppURLEnvKey = "WALLET_APP_URL"
-
 	oidcBasePath    = "/oidc/"
 	healthCheckPath = "/healthcheck"
 	deviceBasePath  = "/device/"
@@ -225,7 +220,6 @@ type httpServerParameters struct {
 	hubAuthURL           string
 	agentUIURL           string
 	logLevel             string
-	walletAppURL         string
 	agent                *agentParameters
 }
 
@@ -334,11 +328,6 @@ func createStartCmd(srv server) *cobra.Command { //nolint:funlen,gocyclo // no r
 				return err
 			}
 
-			walletAppURL, err := cmdutils.GetUserSetVarFromString(cmd, walletAppURLFlagName, walletAppURLEnvKey, true)
-			if err != nil {
-				return err
-			}
-
 			parameters := &httpServerParameters{
 				dependencyMaxRetries: retries,
 				srv:                  srv,
@@ -352,7 +341,6 @@ func createStartCmd(srv server) *cobra.Command { //nolint:funlen,gocyclo // no r
 				hubAuthURL:           hubAuthURL,
 				agentUIURL:           agentUIURL,
 				logLevel:             logLevel,
-				walletAppURL:         walletAppURL,
 				agent:                agentParams,
 			}
 
@@ -374,7 +362,6 @@ func createFlags(startCmd *cobra.Command) {
 	startCmd.Flags().StringP(keyEDVURLFlagName, "", "", keyEDVURLFlagUsage)
 	startCmd.Flags().StringP(userEDVURLFlagName, "", "", userEDVURLFlagUsage)
 	startCmd.Flags().StringP(hubAuthURLFlagName, "", "", hubAuthURLFlagUsage)
-	startCmd.Flags().StringP(walletAppURLFlagName, "", "", walletAppURLFlagUsage)
 
 	createOIDCFlags(startCmd)
 	createTLSFlags(startCmd)
@@ -689,8 +676,7 @@ func router(config *httpServerParameters) (http.Handler, error) {
 
 	// wallet agent router
 	walletHandlers, err := wallet.GetRESTHandlers(ctx, wallet.WithWebhookURLs(config.agent.webhookURLs...),
-		wallet.WithDefaultLabel(config.agent.defaultLabel), wallet.WithMessageHandler(config.agent.msgHandler),
-		wallet.WithWalletAppURL(config.walletAppURL))
+		wallet.WithDefaultLabel(config.agent.defaultLabel), wallet.WithMessageHandler(config.agent.msgHandler))
 	if err != nil {
 		return nil, fmt.Errorf("failed to load wallet handlers: %w", err)
 	}
